Create four_fold_RF
This is a four fold validation version of the RF model. It also retrieves oob error and creates a confusion matrix based on oob error. This is not normal practice, but it is what Velez et al. did, and it allows me to compare my results to theirs since we worked with the same data. Feature importances were also checked in a preliminary way. So far, NDRE results indicate that substituting it for NDVI was an impactful decision. The simplified LAI calculation may be reducing accuracy, as well as Python RF estimators compared to R.